Moved terminal to its own folder
