+++ src/Application.h
@@ -25,54 +25,61 @@
 #include "spdlog/logger.h"
 
 #include "UIWindows/UIWindow.h"
-#include "UIWindows/TerminalWindow.h"
+#include "Terminal/TerminalWindow.h"
 
 #include "UIWindows/Themes/ThemeManager.h"
 #include "Rendering/Texture.h"
 
 
-namespace KBTools{
-class Application {
-public:
-    Application();
+namespace KBTools {
+    class Application {
+    public:
+        Application();
 
-    ~Application();
+        ~Application();
 
-    void AddWindow(const std::string &name, std::shared_ptr<UIWindow> uiWindow) {
-        m_uiWindows[name] = uiWindow;
-    }
+        void AddWindow(const std::string &name, std::shared_ptr<UIWindow> uiWindow) {
+            m_uiWindows[name] = uiWindow;
+        }
 
-    void RemoveWindow(const std::string &name) {
-        m_uiWindows.erase(name);
-    }
+        void RemoveWindow(const std::string &name) {
+            m_uiWindows.erase(name);
+        }
 
 
-private:
-    void InitSDL();
-    void InitImGui();
-    void Run();
+    private:
+        void InitSDL();
 
-    void RenderUI();
-    void SetupDockspace();
+        void InitImGui();
 
-    void Render();
-    GLuint SDLSurfaceToOpenGLTexture(SDL_Surface* surface);
-public:
-    static SDL_Window* GetWindow() { return s_window; }
-private:
-    std::map<std::string, std::shared_ptr<UIWindow>> m_uiWindows;
-    inline static SDL_Window* s_window;
+        void Run();
 
-    SDL_GLContext m_gl_context{};
-    bool m_exit = false;
+        void RenderUI();
 
-    GLsizei m_windowWidth = 1920;
-    GLsizei m_windowHeight = 1080;
+        void SetupDockspace();
 
-    Texture m_texture = Texture("Assets/Textures/newbg.png");
+        void Render();
 
-};
+        GLuint SDLSurfaceToOpenGLTexture(SDL_Surface *surface);
+
+    public:
+        static SDL_Window *GetWindow() { return s_window; }
+
+    private:
+        std::map<std::string, std::shared_ptr<UIWindow>> m_uiWindows;
+        inline static SDL_Window *s_window;
+
+        SDL_GLContext m_gl_context{};
+        bool m_exit = false;
+
+        GLsizei m_windowWidth = 1920;
+        GLsizei m_windowHeight = 1080;
+
+        Texture m_texture = Texture("Assets/Textures/newbg.png");
+
+    };
 
 
 #endif //KBTOOLS_APPLICATION_H
-}
+
+}// namespace KBTools
